Fix stuttering/regression issue caused by "Delaying reads after a seek"
While this commit fixed a bunch of games (CTR's intro music
no longer cuts off early, FF8 Lunar Cry FMV no longer freezes, Worms Pinball
boots properly), it caused a bunch of regressions on games that use ADPCM tracks,
the most noticeable of which was Castlevania Symphony of the Night.

When Dracula is supposed to say "It was not by my hand",
it comes out as "It-it was not by my hand" which is the stutter i'm talking about.
This also happens in other instances in the game, for example when the menu
music plays, there is still a noticeable stutter when it first starts.

What i'm doing here is simply only revert back to the original behaviour
upon an ADPCM track being played. This fixes the issue in Castlevania
but does not seem to regress the other games that were fixed.

I guess though, we should consider this a hack for now.